First batch of changes to permit VS2017 ARM64 tools to build required elements.

Related work items: #14268301

+++ ArmPkg/Drivers/ArmGic/ArmGicLib.c
@@ -131,8 +131,8 @@ ArmGicSendSgiTo (
 {
   MmioWrite32 (
     GicDistributorBase + ARM_GIC_ICDSGIR,
-    ((TargetListFilter & 0x3) << 24) | ((CPUTargetList & 0xFF) << 16) | SgiId
-    );
+	(UINT32)(((TargetListFilter & 0x3) << 24) | ((CPUTargetList & 0xFF) << 16) | SgiId)
+	);  // MS_CHANGE
 }
 
 /*
@@ -213,7 +213,7 @@ ArmGicEnableInterrupt (
   UINTN                 GicCpuRedistributorBase;
 
   // Calculate enable register offset and bit position
-  RegOffset = Source / 32;
+  RegOffset = (UINT32)(Source / 32);    // MS_CHANGE
   RegShift = Source % 32;
 
   Revision = ArmGicGetSupportedArchRevision ();
@@ -257,7 +257,7 @@ ArmGicDisableInterrupt (
   UINTN                 GicCpuRedistributorBase;
 
   // Calculate enable register offset and bit position
-  RegOffset = Source / 32;
+  RegOffset = (UINT32)(Source / 32);   // MS_CHANGE
   RegShift = Source % 32;
 
   Revision = ArmGicGetSupportedArchRevision ();
@@ -301,7 +301,7 @@ ArmGicIsInterruptEnabled (
   UINT32                Interrupts;
 
   // Calculate enable register offset and bit position
-  RegOffset = Source / 32;
+  RegOffset = (UINT32)(Source / 32);  // MS_CHANGE
   RegShift = Source % 32;
 
   Revision = ArmGicGetSupportedArchRevision ();
